Move billing into its own directory.
