+++ src/billing/Members/Members.tsx
@@ -6,7 +6,7 @@ import { EditMemberModal } from 'src/groups/Members/EditMemberModal';
 import { Member, MemberTable } from 'src/groups/Members/MemberTable';
 import { NewMemberModal } from 'src/groups/Members/NewMemberModal';
 import { Ajax } from 'src/libs/ajax';
-import { BillingRole } from 'src/libs/ajax/Billing';
+import { BillingRole } from 'src/libs/ajax/billing/Billing';
 
 interface MembersProps {
   billingProjectName: string;
