+++ src/billing/Workspaces/Workspaces.tsx
@@ -9,7 +9,7 @@ import {
   GoogleBillingAccount,
   isAzureBillingProject,
   isGoogleBillingProject,
-} from 'src/libs/ajax/Billing';
+} from 'src/libs/ajax/billing/Billing';
 import colors from 'src/libs/colors';
 import Events, { extractBillingDetails } from 'src/libs/events';
 import * as Nav from 'src/libs/nav';
