+++ src/workspaces/NewWorkspaceModal/CloneEgressWarning.tsx
@@ -2,7 +2,7 @@ import { Icon, Link } from '@terra-ui-packages/components';
 import { CSSProperties, ReactNode } from 'react';
 import * as React from 'react';
 import { getLocationType, getRegionInfo } from 'src/components/region-common';
-import { BillingProject } from 'src/libs/ajax/Billing';
+import { BillingProject } from 'src/libs/ajax/billing/Billing';
 import { getRegionLabel } from 'src/libs/azure-utils';
 import colors from 'src/libs/colors';
 import * as Utils from 'src/libs/utils';
